GRPC-PYTHON-SERVER/gRPCServer/requirements.txt
cryptography~=43.0.1
pip~=24.3.1
attrs~=24.2.0
retry~=0.9.2
protobuf~=6.33.2
pillow~=11.0.0
filelock>=3.12.1
grpcio~=1.76.0
grpcio-tools~=1.76.0
setuptools~=75.1.0
packaging~=24.1
Cython~=3.0.12
Pygments~=2.18.0
SQLAlchemy~=2.0.45
PyMySQL~=1.1.2
PyJWT~=2.10.1